+++ src/index.ts
@@ -38,6 +38,11 @@ const argv = yargs
         type: "string",
         coerce: coerceJson
     })
+    .option("orgUnitId", {
+        describe:
+            "The target magda org unit id. When provided, all new users will be assigned to this org unit",
+        type: "string"
+    })
     .option("externalUrl", {
         describe: "The base external URL of the gateway.",
         type: "string",
@@ -204,7 +209,8 @@ const authApiClient = new AuthApiClient(
             maxClockSkew: argv?.maxClockSkew,
             allowedExternalRedirectDomains,
             disableLogoutEndpoint: argv?.disableLogoutEndpoint,
-            sessionCookieOptions: argv.cookieJson as any
+            sessionCookieOptions: argv.cookieJson as any,
+            orgUnitId: argv?.orgUnitId
         });
         app.use(routes);