+++ CHANGES.md
@@ -1,6 +1,7 @@
 # v2.0.1
 
 - add OpenID Connect RP-Initiated Logout 1.0 support
+- allow specify the default Magda orgUnit id that the user should be assigned to when logged in for the first time
 
 # v2.0.0